requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose
requests
solana
pytest
httpx